% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandview-package.R
\docType{package}
\name{grandview}
\alias{grandview}
\alias{grandview-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/grandview}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/grandview/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
