% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandview-package.R
\docType{package}
\name{grandview}
\alias{grandview}
\alias{grandview-package}
\title{Provides Dedicated, Platform-Specific Viewers for 'httpgd' Graphics Devices}
\description{
The 'httpgd' packages provides a super-fast, asynchronous graphics device for R graphics
via 'svglite' and 'later' packages. Tools are provided to use dedicated, platform-specific
graphics "browsers" for 'httpgd' graphics devices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/grandview}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/grandview/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
